feat: multi-fault injection list, UI consistency, determinism fixes
Multi-fault system:
- Replace single fault scenario with multi-fault list (FaultList + FaultItem)
- Smart merging: same-tick bursts sum %, same-tick zones take max duration
- Wear-based and Intermittent limited to one per run (UI enforces)
- FaultList::compile() produces FaultConfig + FaultSchedule from list
- New bridge command set_fault_list replaces per-field scenario params
- Remove "Enable fault injection" toggle (implicit from list)
- 7 new compile tests (merging, combined types, empty list)

UI fixes:
- Consistent overlay backgrounds (--bg-overlay variable, all overlays match)
- Canvas background: white in light mode, dark in dark mode
- Obstacle color darkened (0.45 grey) for visibility on white background
- Fault-placed obstacles keep dark brown color after rewind
- Fixed badge widths in fault list (min-width: 48px, centered text)
- Timeline popup: DELETE button replaced with SEEK
- Throughput delta hidden when 0 agents affected

Determinism:
- QueueManager.rebuild_from_agents() preserves queue state on rewind
- operational_age added to snapshots for correct Weibull wear rollback
- FaultConfig default changed to enabled:false, weibull_enabled:false
- 3 new DELETE determinism tests (runner-level parity verified)
- Removed DELETE fault feature (over-engineering, undermined trust)

+++ src/experiment/runner.rs
@@ -484,7 +484,7 @@ fn parse_config_json(json: &str) -> Option<ExperimentConfig> {
             "intermittent" => Some(FaultScenario {
                 enabled: true,
                 scenario_type: FaultScenarioType::IntermittentFault,
-                intermittent_mtbf_ticks: s.get("mtbf").and_then(|t| t.as_u64()).unwrap_or(80) as u32,
+                intermittent_mtbf_ticks: s.get("mtbf").and_then(|t| t.as_u64()).unwrap_or(80),
                 intermittent_recovery_ticks: s.get("recovery").and_then(|r| r.as_u64()).unwrap_or(15) as u32,
                 ..Default::default()
             }),
